backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.22
python-jose[cryptography]>=3.3.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9

# optional helpers
httpx>=0.27.0
yt-dlp>=2025.1.0
